+++ CHANGELOG.md
@@ -16,6 +16,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - Quacc recipes are now decorated with generic `@job`, `@subflow`, or `@flow` decorators to be workflow engine-agnostic.
 - Covalent is now an optional dependency to allow the base `quacc` package to be dependency-light.
 - Default VASP pseudopotentials changed to `setups_pbe54.yaml`.
+- Changed `quacc.recipes.emt.jobflow` and `quacc.recipes.emt.prefect` to `quacc.recipes.emt._jobflow` and `quacc.recipes.emt._prefect`
+- Changed the `quacc[parsl]` extras to install `parsl[monitoring]` instead of just `parsl`
 
 ### Fixed
 
@@ -24,6 +26,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ### Removed
 
+- Removed `quacc.recipes.emt.parsl` since it is now obsolete.
 - Removed `quacc config` command-line option.
 
 ## [0.1.3]